[LargeTensor] Fix int32 overflow in memory_efficient_attention, conv_transpose, and class_center_sample kernels
Fix multiple int32 overflow issues in large tensor scenarios:

1. memory_efficient_attention backward kernel: stride fields (q_strideM, k_strideM,
   v_strideM, gO_strideM, gB_strideM, bias_strideM) and stride methods (o_strideM,
   gQ_strideM, gK_strideM, gV_strideM) changed from int32_t to int64_t to prevent
   overflow in pointer arithmetic when Q tensor numel exceeds INT32_MAX.

2. conv_transpose / conv kernels: group_offset variables changed from int to int64_t
   across forward, backward, and double-grad kernels to prevent overflow for large
   feature maps. Added output numel check in ConvTransposeInferMeta.

3. class_center_sample kernel: added int32 numel check in InferMeta; fixed implicit
   int64->int narrowing in cub sort call.

4. im2col.cu / vol2col.cu: changed intermediate channel variables from int to int64_t
   to prevent overflow in col2im and vol2col kernels.

+++ paddle/phi/kernels/funcs/im2col.cu
@@ -184,7 +184,7 @@ __global__ void col2im(int64_t n,
   const int d_filter_width = dilation_w * (filter_width - 1) + 1;
 
   // NOTE(zrr1999): input_channels must be less than the range of int32
-  int input_channels = n / im_height / im_width;
+  int64_t input_channels = n / im_height / im_width;
 
   if (index < n) {
     T val = static_cast<T>(0);
@@ -195,8 +195,8 @@ __global__ void col2im(int64_t n,
                      ? (index / im_width) % im_height + padding_height
                      : (index / input_channels / im_width) % im_height +
                            padding_height);
-    int c = (data_layout != DataLayout::NHWC ? index / im_width / im_height
-                                             : index % input_channels);
+    int64_t c = (data_layout != DataLayout::NHWC ? index / im_width / im_height
+                                                 : index % input_channels);
 
     // compute the start and end of the output
     int64_t w_col_start =
