+++ paddle/phi/kernels/funcs/vol2col.cu
@@ -46,7 +46,7 @@ __global__ void vol2col(int64_t num_kernels,
                         int64_t output_width,
                         T* data_col,
                         const DataLayout data_layout) {
-  int input_channels =
+  int64_t input_channels =
       num_kernels / output_detph / output_height / output_width;
   for (int64_t index =
            static_cast<int64_t>(blockIdx.x) * blockDim.x + threadIdx.x;
@@ -55,8 +55,9 @@ __global__ void vol2col(int64_t num_kernels,
     int64_t w_out = index % output_width;
     int64_t h_out = (index / output_width) % output_height;
     int64_t d_out = (index / output_width / output_height) % output_detph;
-    int channel_in = index / output_width / output_height / output_detph;
-    int channel_out = channel_in * filter_depth * filter_height * filter_width;
+    int64_t channel_in = index / output_width / output_height / output_detph;
+    int64_t channel_out =
+        channel_in * filter_depth * filter_height * filter_width;
     int64_t w_in = w_out * stride_width - padding_width;
     int64_t h_in = h_out * stride_height - padding_height;
     int64_t d_in = d_out * stride_depth - padding_depth;
